backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.121.2
h11==0.16.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
pyairtable==3.3.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
google-cloud-firestore>=2.11.0
google-cloud-logging>=3.8.0
google-cloud-secret-manager>=2.16.0
tenacity>=8.2.3
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0
firebase-admin>=6.0.0
